Conflicts:
	.travis.yml
	README.md
	build.properties
	gradle/forge.gradle
	release.sh
	src/main/java/evilcraft/EvilCraft.java
	src/main/java/evilcraft/Recipes.java
	src/main/java/evilcraft/api/ILocation.java
	src/main/java/evilcraft/api/RegistryManager.java
	src/main/java/evilcraft/block/BloodInfuser.java
	src/main/java/evilcraft/block/DarkBloodBrick.java
	src/main/java/evilcraft/block/EnvironmentalAccumulator.java
	src/main/java/evilcraft/block/SpiritFurnace.java
	src/main/java/evilcraft/client/gui/container/GuiOriginsOfDarkness.java
	src/main/java/evilcraft/core/PlayerExtendedInventoryIterator.java
	src/main/java/evilcraft/core/PlayerInventoryIterator.java
	src/main/java/evilcraft/core/algorithm/Location.java
	src/main/java/evilcraft/core/client/render/tileentity/RenderTileEntityModel.java
	src/main/java/evilcraft/core/config/ConfigProperty.java
	src/main/java/evilcraft/core/config/configurable/ConfigurableVillager.java
	src/main/java/evilcraft/core/config/configurabletypeaction/MobAction.java
	src/main/java/evilcraft/core/config/extendedconfig/ItemConfig.java
	src/main/java/evilcraft/core/config/extendedconfig/MobConfig.java
	src/main/java/evilcraft/core/helper/EntityHelpers.java
	src/main/java/evilcraft/core/helper/LocationHelpers.java
	src/main/java/evilcraft/core/helper/RenderHelpers.java
	src/main/java/evilcraft/core/inventory/NBTSimpleInventoryItemHeld.java
	src/main/java/evilcraft/core/inventory/container/InventoryContainer.java
	src/main/java/evilcraft/core/inventory/container/ItemInventoryContainer.java
	src/main/java/evilcraft/core/inventory/slot/SlotFluidContainer.java
	src/main/java/evilcraft/core/item/DamageIndicatedItemComponent.java
	src/main/java/evilcraft/core/tileentity/TankInventoryTileEntity.java
	src/main/java/evilcraft/infobook/InfoBookParser.java
	src/main/java/evilcraft/item/DarkStick.java
	src/main/java/evilcraft/modcompat/ModCompatLoader.java
	src/main/java/evilcraft/network/packet/SoundPacket.java
	src/main/java/evilcraft/world/gen/DarkTempleGenerator.java
	src/main/java/evilcraft/world/gen/nbt/DarkTempleData.java
	src/main/java/org/cyclops/evilcraft/Configs.java
	src/main/java/org/cyclops/evilcraft/GeneralConfig.java
	src/main/java/org/cyclops/evilcraft/Reference.java
	src/main/java/org/cyclops/evilcraft/block/BoxOfEternalClosure.java
	src/main/java/org/cyclops/evilcraft/block/BoxOfEternalClosureConfig.java
	src/main/java/org/cyclops/evilcraft/block/SpiritFurnaceConfig.java
	src/main/java/org/cyclops/evilcraft/client/gui/container/GuiBloodChest.java
	src/main/java/org/cyclops/evilcraft/client/particle/EntityBloodBubbleFX.java
	src/main/java/org/cyclops/evilcraft/client/particle/EntityBlurFX.java
	src/main/java/org/cyclops/evilcraft/client/render/entity/RenderVengeanceSpirit.java
	src/main/java/org/cyclops/evilcraft/client/render/tileentity/RenderTileEntityPurifier.java
	src/main/java/org/cyclops/evilcraft/core/block/CubeDetector.java
	src/main/java/org/cyclops/evilcraft/core/block/HollowCubeDetector.java
	src/main/java/org/cyclops/evilcraft/core/client/gui/container/GuiContainerTankInventory.java
	src/main/java/org/cyclops/evilcraft/core/degradation/effect/KnockbackDistortDegradation.java
	src/main/java/org/cyclops/evilcraft/core/helper/ItemHelpers.java
	src/main/java/org/cyclops/evilcraft/core/recipe/custom/EnvironmentalAccumulatorRecipeProperties.java
	src/main/java/org/cyclops/evilcraft/core/recipe/xml/BloodInfuserRecipeTypeHandler.java
	src/main/java/org/cyclops/evilcraft/enchantment/EnchantmentUnusing.java
	src/main/java/org/cyclops/evilcraft/entity/effect/EntityNecromancersHead.java
	src/main/java/org/cyclops/evilcraft/entity/monster/VengeanceSpirit.java
	src/main/java/org/cyclops/evilcraft/entity/monster/VengeanceSpiritConfig.java
	src/main/java/org/cyclops/evilcraft/entity/villager/WerewolfVillager.java
	src/main/java/org/cyclops/evilcraft/infobook/pageelement/BloodInfuserRecipeAppendix.java
	src/main/java/org/cyclops/evilcraft/infobook/pageelement/EnvironmentalAccumulatorRecipeAppendix.java
	src/main/java/org/cyclops/evilcraft/inventory/container/ContainerBloodChest.java
	src/main/java/org/cyclops/evilcraft/inventory/container/ContainerBloodInfuser.java
	src/main/java/org/cyclops/evilcraft/item/BloodExtractor.java
	src/main/java/org/cyclops/evilcraft/item/BroomConfig.java
	src/main/java/org/cyclops/evilcraft/item/DarkStickConfig.java
	src/main/java/org/cyclops/evilcraft/item/ExaltedCrafter.java
	src/main/java/org/cyclops/evilcraft/item/Kineticator.java
	src/main/java/org/cyclops/evilcraft/item/MaceOfDistortion.java
	src/main/java/org/cyclops/evilcraft/item/OriginsOfDarknessConfig.java
	src/main/java/org/cyclops/evilcraft/item/WerewolfFleshConfig.java
	src/main/java/org/cyclops/evilcraft/modcompat/nei/ExaltedCrafterOverlayHandler.java
	src/main/java/org/cyclops/evilcraft/modcompat/nei/NEIBloodInfuserManager.java
	src/main/java/org/cyclops/evilcraft/modcompat/nei/NEIEnvironmentalAccumulatorManager.java
	src/main/java/org/cyclops/evilcraft/modcompat/nei/NEIEvilCraftConfig.java
	src/main/java/org/cyclops/evilcraft/modcompat/waila/TankDataProvider.java
	src/main/java/org/cyclops/evilcraft/tileentity/TileBloodInfuser.java
	src/main/java/org/cyclops/evilcraft/tileentity/TileDarkTank.java
	src/main/java/org/cyclops/evilcraft/tileentity/TileEnvironmentalAccumulator.java
	src/main/java/org/cyclops/evilcraft/tileentity/TilePurifier.java
	src/main/java/org/cyclops/evilcraft/tileentity/tickaction/EmptyItemBucketInTankTickAction.java
	src/main/java/org/cyclops/evilcraft/tileentity/tickaction/bloodchest/BloodChestRepairActionRegistry.java
	src/main/java/org/cyclops/evilcraft/tileentity/tickaction/bloodchest/RepairItemTickAction.java
	src/main/java/org/cyclops/evilcraft/tileentity/tickaction/bloodinfuser/InfuseItemTickAction.java
	src/main/java/org/cyclops/evilcraft/tileentity/tickaction/spiritfurnace/BoxCookTickAction.java
	src/main/java/org/cyclops/evilcraft/world/gen/structure/DarkTempleStructure.java
	src/main/java/org/cyclops/evilcraft/world/gen/structure/QuarterSymmetricalStructure.java
	src/main/resources/assets/evilcraft/info/book.xml
	src/main/resources/assets/evilcraft/lang/en_US.lang
	src/main/resources/assets/evilcraft/lang/zh_CN.lang
	src/todo/java/evilcraft/modcompat/bloodmagic/BloodMagicModCompat.java
	src/todo/java/evilcraft/modcompat/thermalexpansion/ThermalExpansionModCompat.java

+++ src/main/java/org/cyclops/evilcraft/core/tileentity/WorkingTileEntity.java
@@ -100,7 +100,7 @@ public abstract class WorkingTileEntity<T extends TankInventoryTileEntity, O> ex
         return getTickers().get(getWorkTicker()).getTick();
     }
     
-    protected int getRequiredWorkTicks() {
+    protected float getRequiredWorkTicks() {
         return getTickers().get(getWorkTicker()).getRequiredTicks();
     }
